% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{barbell_weights}
\alias{barbell_weights}
\title{Calculate the weights of a duration matched barbell hedging portfolio}
\usage{
barbell_weights(duration_bullet, duration_asset_1, duration_asset_2)
}
\arguments{
\item{duration_bullet}{The bullet duration}

\item{duration_asset_1}{The first hedging asset's duration}

\item{duration_asset_2}{The second hedging asset's duration}
}
\value{
The weight of asset_1. The weight of asset_2 is simply (1 - weight_asset_1).
}
\description{
Using the duration of a bullet and the durations of two assets, this function
calculates the weighting necessary to hedge the risk in the bullet, assuming
a flat yield curve. The durations can be Modified or Macaulay.
}
\details{
Only the weight of asset_1 is returned so that this function can be called
easily from a \code{dplyr} workflow.
}
