% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_rates_xls}
\alias{download_rates_xls}
\title{Download the rates data from the Federal Reserve}
\usage{

  download_rates_xls(url = "https://www.federalreserve.gov/econresdata/researchdata/feds200628.xls",
  dest = "rate_data.xls", overwrite = FALSE)
}
\description{
Download the rates data from the Federal Reserve
}
