% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration}
\alias{duration}
\title{Calculate the duration of a bond}
\usage{
duration(n, c, M, y, type = "modified")
}
\arguments{
\item{n}{The number of 6 month \emph{periods} in the bond.}

\item{c}{The annualized semi-annual coupon rate, as a decimal.}

\item{M}{The maturity value of the bond.}

\item{y}{The yield to maturity on the bond, as a decimal.}

\item{type}{Either \code{"modified"} or \code{"macaulay"}.}
}
\description{
Calculate the duration of a bond
}
