% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{list_named}
\alias{list_named}
\title{Create a named list using the quoted names of the list components}
\usage{
list_named(...)
}
\description{
Create a named list using the quoted names of the list components
}
