% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price-bond.R
\name{price_bond}
\alias{price_bond}
\title{Price a coupon bond}
\usage{
price_bond(n, c, M, y)
}
\arguments{
\item{n}{The number of 6 month \emph{periods} in the bond.}

\item{c}{The annualized semi-annual coupon rate, as a decimal.}

\item{M}{The maturity value of the bond.}

\item{y}{The yield to maturity on the bond, as a decimal.}
}
\description{
Price an n-period semi-annual coupon bond.
}
\details{
Any of these variables can be a vector, and as long as the other variables
are either the same length, or length 1, you can calculate multiple bond
prices at once.
}
