% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_rates}
\alias{read_rates}
\alias{read_rates_meta_data}
\title{Read in only the rates data}
\usage{
read_rates(path, sheet = NULL, range = NULL, col_names = TRUE,
  col_types = NULL, na = c("", "-999.99"), trim_ws = TRUE, skip = 9,
  n_max = Inf, guess_max = min(1000, n_max), clean_date_col = TRUE)

read_rates_meta_data(path, sheet = NULL, range = "A3:H8",
  col_names = TRUE, col_types = NULL, na = c("", "N/A"), trim_ws = TRUE,
  skip = 0, n_max = Inf, guess_max = min(1000, n_max))
}
\description{
Functions for reading rates data, with sane defaults for NA handling. \code{\link[=read_rates]{read_rates()}}
reads only the interest rates data, and \code{\link[=read_rates_meta_data]{read_rates_meta_data()}} reads in
only the meta data.
}
