% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-rates.R
\name{spot_rate_factory}
\alias{spot_rate_factory}
\title{Create the Gurkaynak, Sack, and Wright spot rate function}
\usage{
spot_rate_factory(beta_0, beta_1, beta_2, beta_3, tau_1, tau_2)
}
\description{
Create the spot rate function y_t(n) using equation 22 of
Gurkaynak, Sack, and Wright (2006). The returned function is
parameterized by \code{n}, corresponding to the n-year continuously
compounded spot rates at any given date t.
}
