% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot-rates.R
\name{zero_bond_price_factory}
\alias{zero_bond_price_factory}
\title{Calculate zero coupon bond prices from spot rates}
\usage{
zero_bond_price_factory(spot_rate_fn)
}
\description{
Given a spot rate function taking a single argument, \code{n}, create a function for calculating
n-year zero coupon bond prices.
}
